+++ src/main/java/com/za/minecraft/network/packets/BlockUpdatePacket.java
@@ -2,13 +2,13 @@ package com.za.minecraft.network.packets;
 
 public class BlockUpdatePacket extends NetworkPacket {
     public int x, y, z;
-    public byte blockType;
+    public int blockType;
     public long timestamp;
     
     public BlockUpdatePacket() {
     }
     
-    public BlockUpdatePacket(int x, int y, int z, byte blockType) {
+    public BlockUpdatePacket(int x, int y, int z, int blockType) {
         this.x = x;
         this.y = y;
         this.z = z;
